[data.search.SearchSource] Move some SearchSource dependencies to the server. (#74607)
