--- docs/development/plugins/data/public/kibana-plugin-plugins-data-public.getsearcherrortype.md
@@ -1,22 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [kibana-plugin-plugins-data-public](./kibana-plugin-plugins-data-public.md) &gt; [getSearchErrorType](./kibana-plugin-plugins-data-public.getsearcherrortype.md)
-
-## getSearchErrorType() function
-
-<b>Signature:</b>
-
-```typescript
-export declare function getSearchErrorType({ message }: Pick<SearchError, 'message'>): "UNSUPPORTED_QUERY" | undefined;
-```
-
-## Parameters
-
-|  Parameter | Type | Description |
-|  --- | --- | --- |
-|  { message } | <code>Pick&lt;SearchError, 'message'&gt;</code> |  |
-
-<b>Returns:</b>
-
-`"UNSUPPORTED_QUERY" | undefined`
-
