--- docs/development/plugins/data/public/kibana-plugin-plugins-data-public.searcherror._constructor_.md
@@ -1,20 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [kibana-plugin-plugins-data-public](./kibana-plugin-plugins-data-public.md) &gt; [SearchError](./kibana-plugin-plugins-data-public.searcherror.md) &gt; [(constructor)](./kibana-plugin-plugins-data-public.searcherror._constructor_.md)
-
-## SearchError.(constructor)
-
-Constructs a new instance of the `SearchError` class
-
-<b>Signature:</b>
-
-```typescript
-constructor({ status, title, message, path, type }: SearchErrorOptions);
-```
-
-## Parameters
-
-|  Parameter | Type | Description |
-|  --- | --- | --- |
-|  { status, title, message, path, type } | <code>SearchErrorOptions</code> |  |
-
