+++ src/plugins/data/public/index.ts
@@ -337,7 +337,6 @@ export {
   // search
   ES_SEARCH_STRATEGY,
   getEsPreference,
-  getSearchErrorType,
   ISearch,
   ISearchOptions,
   ISearchGeneric,
