+++ src/plugins/data/public/search/fetch/index.ts
@@ -27,6 +27,5 @@ export {
   getMaxConcurrentShardRequests,
 } from './get_search_params';
 
-export { SearchError, getSearchErrorType } from './search_error';
 export { RequestFailure } from './request_error';
 export { handleResponse } from './handle_response';
