+++ src/plugins/data/public/search/index.ts
@@ -34,7 +34,6 @@ export {
   FetchOptions,
   SearchRequest,
   SearchResponse,
-  getSearchErrorType,
   getSearchParamsFromRequest,
 } from './fetch';
