+++ src/plugins/data/public/search/search_source/search_source.ts
@@ -75,7 +75,7 @@ import { map } from 'rxjs/operators';
 import { CoreStart } from 'kibana/public';
 import { normalizeSortRequest } from './normalize_sort_request';
 import { filterDocvalueFields } from './filter_docvalue_fields';
-import { fieldWildcardFilter } from '../../../../kibana_utils/public';
+import { fieldWildcardFilter } from '../../../../kibana_utils/common';
 import { IIndexPattern, ISearchGeneric, SearchRequest } from '../..';
 import { SearchSourceOptions, SearchSourceFields } from './types';
 import { FetchOptions, RequestFailure, handleResponse, getSearchParamsFromRequest } from '../fetch';
