+++ src/plugins/kibana_legacy/common/migrate_legacy_query.ts
