+++ src/plugins/kibana_legacy/public/index.ts
@@ -24,7 +24,7 @@ export const plugin = (initializerContext: PluginInitializerContext) =>
   new KibanaLegacyPlugin(initializerContext);
 
 export * from './plugin';
-export { kbnBaseUrl } from '../common/kbn_base_url';
+export { kbnBaseUrl, migrateLegacyQuery } from '../common';
 
 export { initAngularBootstrap } from './angular_bootstrap';
 export { PaginateDirectiveProvider, PaginateControlsDirectiveProvider } from './paginate/paginate';
