+++ src/plugins/kibana_utils/common/field_wildcard.ts
@@ -19,6 +19,7 @@
 
 import { escapeRegExp, memoize } from 'lodash';
 
+// @internal
 export const makeRegEx = memoize(function makeRegEx(glob: string) {
   const globRegex = glob.split('*').map(escapeRegExp).join('.*');
   return new RegExp(`^${globRegex}$`);
