+++ src/plugins/kibana_utils/public/index.ts
@@ -21,6 +21,8 @@ export {
   calculateObjectHash,
   defer,
   Defer,
+  fieldWildcardFilter,
+  fieldWildcardMatcher,
   Get,
   JsonArray,
   JsonObject,
@@ -34,7 +36,6 @@ export {
 } from '../common';
 export * from './core';
 export * from '../common/errors';
-export * from './field_wildcard';
 export * from './render_complete';
 export * from './resize_checker';
 export * from '../common/state_containers';
